Folded the unit formatting checks into FullScanPropertiesTest
* Moved the hand-built term cases from OtherMetadataInfoTest into a VerifyUnitFormatting method on the test that already covers the Other Metadata display, and removed the separate test class

See ai/todos/active/TODO-20260630_mzml_cv_metadata.md
